4_personal_agent/requirements.txt
flask
openai
python-dotenv
httpx
google-auth-oauthlib
google-api-python-client
beautifulsoup4
requests
